Feature update (#305)
-Removes VG modes from MODE_ORIGINAL
-Adds rotary fire mode selector inspired by Kaden
-Adds cyclotron slime animation inspired by Kaden
-Adds remaining VG alt fire modes
-Adds long press quick vent to perform full vent
-Enables quick vent function by default
-Removes "venting" fire mode
-Removes "settings" fire mode
-Adds ribbon cable attach/detach sounds
-Updates FE boot effect sound
-Adds "thunk" switch sound to ion arm switch regardless of operation
-Fixes Frutto barrel LED animations
-Fixes missing 1989 mood slime color on Attenuator
-Program size and RAM use reduction
-Various datatype fixes
-Documentation updates

+++ source/Attenuator/Colours.h
@@ -46,19 +46,19 @@ enum colours {
   C_SPECTRAL_CUSTOM
 };
 
-int getBrightness(uint8_t i_percent = 100) {
+uint8_t getBrightness(uint8_t i_percent = 100) {
   // Brightness here is a percentage, to be converted to a range 0-255.
   if(i_percent > 100) {
     i_percent = 100;
   }
-  return (int) ((255 * i_percent) / 100);
+  return (uint8_t) ((255 * i_percent) / 100);
 }
 
 // Special values for colour cycles: current hue (colour) and when to change colour.
 // This must match the number of device ENUM entries (though that is rarely changed).
 uint8_t i_curr_colour[DEVICE_NUM_LEDS] = { 0, 0, 0 };
 uint8_t i_curr_bright[DEVICE_NUM_LEDS] = { 0, 0, 0 };
-int i_next_bright[DEVICE_NUM_LEDS] = { -1, -1, -1 }; // Uses int to allow negative steps.
+int16_t i_next_bright[DEVICE_NUM_LEDS] = { -1, -1, -1 }; // Uses int to allow negative steps.
 uint8_t i_count[DEVICE_NUM_LEDS] = { 0, 0, 0 }; // Counter-based changes for certain themes.
 millisDelay ms_color_change[DEVICE_NUM_LEDS]; // Timers for changing colors for certain themes.
 uint16_t i_change_delay[DEVICE_NUM_LEDS] = { 10, 10, 10 }; // Default delay time for changes.
